Makefile:
For MSWindows, the automatic build on github was written to pass the
name of the executable to make, avoiding the new 'all' target I added
to deal with the dependencies on the 'make depend' files and
version.inc.

This commit also adds $(EXE) based on the OS via uname; if that command
is not available, MSWindows is presumed.